+++ packages/react/src/alert-dialog/index.parts.ts
@@ -1,6 +1,3 @@
-import { DialogHandle } from '../dialog/store/DialogHandle';
-import { DialogStore } from '../dialog/store/DialogStore';
-
 export { AlertDialogRoot as Root } from './root/AlertDialogRoot';
 export { DialogBackdrop as Backdrop } from '../dialog/backdrop/DialogBackdrop';
 export { DialogClose as Close } from '../dialog/close/DialogClose';
@@ -10,13 +7,5 @@ export { DialogPortal as Portal } from '../dialog/portal/DialogPortal';
 export { DialogTitle as Title } from '../dialog/title/DialogTitle';
 export { DialogTrigger as Trigger } from '../dialog/trigger/DialogTrigger';
 export { DialogViewport as Viewport } from '../dialog/viewport/DialogViewport';
-
-export function createHandle<Payload>(): DialogHandle<Payload> {
-  return new DialogHandle<Payload>(
-    new DialogStore<Payload>({
-      modal: true,
-      disablePointerDismissal: true,
-      role: 'alertdialog',
-    }),
-  );
-}
+export { createAlertDialogHandle as createHandle } from './handle';
+export { DialogHandle as Handle } from '../dialog/store/DialogHandle';