+++ packages/react/src/menu/store/MenuStore.ts
@@ -3,9 +3,9 @@ import { createSelector, ReactStore } from '@base-ui-components/utils/store';
 import { EMPTY_OBJECT } from '@base-ui-components/utils/empty';
 import { useRefWithInit } from '@base-ui-components/utils/useRefWithInit';
 import { MenuParent, MenuRoot } from '../root/MenuRoot';
-import { FloatingRootContext } from '../../floating-ui-react';
-import { getEmptyContext } from '../../floating-ui-react/hooks/useFloatingRootContext';
-import { FloatingTreeStore } from '../../floating-ui-react/components/FloatingTree';
+import { type FloatingRootContext } from '../../floating-ui-react';
+import { getEmptyRootContext } from '../../floating-ui-react/utils/getEmptyRootContext';
+import { FloatingTreeStore } from '../../floating-ui-react/components/FloatingTreeStore';
 import { TransitionStatus } from '../../utils/useTransitionStatus';
 import { HTMLProps } from '../../utils/types';
 import { PopupTriggerMap } from '../../utils/popupStoreUtils';
@@ -227,7 +227,7 @@ function createInitialState<Payload>(): State<Payload> {
     transitionStatus: 'idle',
     instantType: undefined,
     lastOpenChangeReason: null,
-    floatingRootContext: getEmptyContext(),
+    floatingRootContext: getEmptyRootContext(),
     floatingTreeRoot: new FloatingTreeStore(),
     floatingNodeId: undefined,
     floatingParentNodeId: null,